refactor(components-react): move visreg scenarios to VisualStates components
