+++ packages/storybook/src/react-components/form-field-description/visual/States.tsx
@@ -1,0 +1,14 @@
+import { LuxFormFieldDescription } from '@lux-design-system/components-react';
+
+export const VisualStates = () => (
+  <>
+    <LuxFormFieldDescription>Form Field Description</LuxFormFieldDescription>
+    <LuxFormFieldDescription appearance="valid">Valid Form Field Description</LuxFormFieldDescription>
+    <LuxFormFieldDescription appearance="invalid">Invalid Form Field Description</LuxFormFieldDescription>
+    <LuxFormFieldDescription>
+      Long Form Field Description. Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Dolor ante id varius,
+      aenean eu faucibus vitae malesuada. Viverra malesuada aliquam et placerat justo porta ipsum parturient. Cursus
+      nostra varius efficitur lobortis aliquam lectus bibendum.
+    </LuxFormFieldDescription>
+  </>
+);
