+++ packages/storybook/src/react-components/form-field-error-message/visual/States.tsx
@@ -1,0 +1,7 @@
+import { LuxFormFieldErrorMessage } from '@lux-design-system/components-react';
+
+export const VisualStates = () => (
+  <>
+    <LuxFormFieldErrorMessage>Error Message</LuxFormFieldErrorMessage>
+  </>
+);
