+++ packages/storybook/src/react-components/form-field-label/visual/States.tsx
@@ -1,0 +1,11 @@
+import { LuxFormFieldLabel } from '@lux-design-system/components-react';
+
+export const VisualStates = () => (
+  <>
+    <LuxFormFieldLabel>Label</LuxFormFieldLabel>
+    <LuxFormFieldLabel type="checkbox">Checkbox Label</LuxFormFieldLabel>
+    <LuxFormFieldLabel type="radio">Radio Label</LuxFormFieldLabel>
+    <LuxFormFieldLabel checked>Checked Label</LuxFormFieldLabel>
+    <LuxFormFieldLabel disabled>Disabled Label</LuxFormFieldLabel>
+  </>
+);
